refactor backup-finish.tsx
