+++ src/i18n.ts
@@ -144,6 +144,10 @@ export enum I18N {
   welcomeCreateWallet,
   welcomeLedgerWallet,
   welcomeRestoreWallet,
+  backupFinishcongrat,
+  backupFinishSuccess,
+  backupFinishFinish
+
 }
 
 export function getText(key: I18N, params?: Record<string, string>): string {
@@ -321,4 +325,7 @@ const en: Record<I18N, string> = {
   [I18N.welcomeCreateWallet]: 'Create Wallet',
   [I18N.welcomeLedgerWallet]: 'Connect',
   [I18N.welcomeRestoreWallet]: 'I already have a wallet',
+  [I18N.backupFinishcongrat]: 'Congratulations!',
+  [I18N.backupFinishSuccess]: "You've successfully protected your wallet.!",
+  [I18N.backupFinishFinish]: "Finish",
 };
